+++ project/FrontEnd/collaborative_science_platform/lib/screens/home_page/desktop_mobile_page.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+class DesktopMobilePage extends StatelessWidget {
+  const DesktopMobilePage({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const Placeholder();
+  }
+}
