+++ project/FrontEnd/collaborative_science_platform/lib/screens/node_details_page/widgets/contributors_list_view.dart
