+++ project/FrontEnd/collaborative_science_platform/lib/screens/page_with_appbar/page_with_appbar.dart
