+++ project/FrontEnd/collaborative_science_platform/lib/screens/page_with_appbar/widgets/app_bar_logo.dart
