+++ project/FrontEnd/collaborative_science_platform/lib/screens/page_with_appbar/widgets/profile_menu.dart
@@ -1,10 +1,9 @@
 import 'package:collaborative_science_platform/providers/auth.dart';
 import 'package:collaborative_science_platform/screens/auth_screens/login_page.dart';
 import 'package:collaborative_science_platform/screens/auth_screens/signup_page.dart';
-import 'package:collaborative_science_platform/screens/node_details_page/node_details_page.dart';
 import 'package:collaborative_science_platform/screens/profile_page/profile_page.dart';
 import 'package:collaborative_science_platform/services/screen_navigation.dart';
-import 'package:collaborative_science_platform/widgets/app_bar_widgets/app_bar_button.dart';
+import 'package:collaborative_science_platform/screens/page_with_appbar/widgets/app_bar_button.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -15,15 +14,12 @@ class ProfileMenu extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final auth = Provider.of<Auth>(context);
-    return auth.isSignedIn
-        ? AuthenticatedProfileMenu()
-        : UnAuthenticatedProfileMenu();
+    return auth.isSignedIn ? AuthenticatedProfileMenu() : UnAuthenticatedProfileMenu();
   }
 }
 
 class AuthenticatedProfileMenu extends StatelessWidget {
-  final GlobalKey<PopupMenuButtonState<dynamic>> _popupMenu =
-      GlobalKey<PopupMenuButtonState>();
+  final GlobalKey<PopupMenuButtonState<dynamic>> _popupMenu = GlobalKey<PopupMenuButtonState>();
   AuthenticatedProfileMenu({super.key});
 
   @override
@@ -36,11 +32,9 @@ class AuthenticatedProfileMenu extends StatelessWidget {
       onSelected: (String result) async {
         switch (result) {
           case 'profile':
-            Provider.of<ScreenNavigation>(context, listen: false)
-                .setSelectedTab(ScreenTab.profile);
-            Navigator.pushNamed(context, ProfilePage.routeName,
-                arguments: auth.user!.email);
-                                    
+            Provider.of<ScreenNavigation>(context, listen: false).setSelectedTab(ScreenTab.profile);
+            Navigator.pushNamed(context, ProfilePage.routeName, arguments: auth.user!.email);
+
             break;
           case 'logout':
             auth.logout();
@@ -68,8 +62,7 @@ class AuthenticatedProfileMenu extends StatelessWidget {
 }
 
 class UnAuthenticatedProfileMenu extends StatelessWidget {
-  final GlobalKey<PopupMenuButtonState<dynamic>> _popupMenu =
-      GlobalKey<PopupMenuButtonState>();
+  final GlobalKey<PopupMenuButtonState<dynamic>> _popupMenu = GlobalKey<PopupMenuButtonState>();
   UnAuthenticatedProfileMenu({super.key});
 
   @override
