+++ project/FrontEnd/collaborative_science_platform/lib/screens/profile_page/change_password_page.dart
@@ -1,10 +1,7 @@
-import 'package:collaborative_science_platform/screens/page_with_appbar.dart';
+import 'package:collaborative_science_platform/screens/page_with_appbar/page_with_appbar.dart';
+import 'package:collaborative_science_platform/screens/profile_page/widgets/change_password_form.dart';
 import 'package:collaborative_science_platform/screens/profile_page/widgets/settings_appbar.dart';
-import 'package:collaborative_science_platform/utils/colors.dart';
 import 'package:flutter/material.dart';
-import 'package:collaborative_science_platform/models/profile_data.dart';
-import 'package:collaborative_science_platform/screens/profile_page/widgets/settings_input_widget.dart';
-import 'package:collaborative_science_platform/utils/responsive/responsive.dart';
 
 class ChangePasswordPage extends StatelessWidget {
   static const routeName = '/change-password';
@@ -19,81 +16,3 @@ class ChangePasswordPage extends StatelessWidget {
   }
 }
 
-class ChangePasswordForm extends StatefulWidget {
-  const ChangePasswordForm({
-    super.key,
-  });
-
-  @override
-  State<ChangePasswordForm> createState() => _ChangePasswordFormState();
-}
-
-class _ChangePasswordFormState extends State<ChangePasswordForm> {
-  ProfileData profileData = ProfileData();
-  final passwordController = TextEditingController();
-  final confirmPasswordController = TextEditingController();
-
-  final passwordFocusNode = FocusNode();
-  final confirmPasswordFocusNode = FocusNode();
-  final double x = 300;
-  @override
-  void dispose() {
-    passwordController.dispose();
-    passwordFocusNode.dispose();
-    confirmPasswordController.dispose();
-    confirmPasswordFocusNode.dispose();
-    super.dispose();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    double screenWidth = MediaQuery.of(context).size.width;
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 36.0),
-      width: Responsive.getGenericPageWidth(context),
-      child: Column(
-        children: [
-          Column(
-            crossAxisAlignment: CrossAxisAlignment.center,
-            children: [
-              const SizedBox(height: 20.0),
-              SettingsWidget(
-                controller: passwordController,
-                focusNode: passwordFocusNode,
-                textType: "New password",
-                prefixIcon: Icons.lock,
-                widgetWidth: screenWidth,
-              ),
-              const SizedBox(height: 20.0),
-              SettingsWidget(
-                controller: confirmPasswordController,
-                focusNode: confirmPasswordFocusNode,
-                textType: "Confirm New password",
-                prefixIcon: Icons.lock,
-                widgetWidth: screenWidth,
-              ),
-            ],
-          ),
-          const SizedBox(height: 20.0),
-          MouseRegion(
-            cursor: SystemMouseCursors.click,
-            child: GestureDetector(
-              onTap: () {},
-              child: Container(
-                height: 40.0,
-                width: MediaQuery.of(context).size.width - 40,
-                decoration: BoxDecoration(color: AppColors.primaryColor, borderRadius: BorderRadius.circular(5.0)),
-                child: const Row(
-                  mainAxisAlignment: MainAxisAlignment.center,
-                  children: [
-                    Text('Save', style: TextStyle(color: Colors.white, fontWeight: FontWeight.bold, fontSize: 16.0)),
-                  ],
-                ),
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
