+++ project/FrontEnd/collaborative_science_platform/lib/services/screen_navigation.dart
@@ -1,9 +1,9 @@
 import 'package:collaborative_science_platform/screens/auth_screens/please_login_page.dart';
 import 'package:collaborative_science_platform/screens/home_page/home_page.dart';
-import 'package:collaborative_science_platform/screens/notifications_page.dart';
+import 'package:collaborative_science_platform/screens/notifications_page/notifications_page.dart';
 import 'package:collaborative_science_platform/screens/graph_page.dart';
 import 'package:collaborative_science_platform/screens/profile_page/profile_page.dart';
-import 'package:collaborative_science_platform/screens/workspaces_page.dart';
+import 'package:collaborative_science_platform/screens/workspaces_page/workspaces_page.dart';
 import 'package:flutter/material.dart';
 
 enum ScreenTab {
